+++ airflow/providers/amazon/aws/sensors/s3_keys_unchanged.py
@@ -19,10 +19,10 @@
 
 import warnings
 
-from airflow.providers.amazon.aws.sensors.s3 import S3KeysUnchangedSensor # noqa
+from airflow.providers.amazon.aws.sensors.s3 import S3KeysUnchangedSensor  # noqa
 
 warnings.warn(
     "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.s3`.",
     DeprecationWarning,
-    stacklevel=2
+    stacklevel=2,
 )
