+++ airflow/providers/amazon/aws/sensors/s3_prefix.py
@@ -19,10 +19,10 @@
 
 import warnings
 
-from airflow.providers.amazon.aws.sensors.s3 import S3PrefixSensor # noqa
+from airflow.providers.amazon.aws.sensors.s3 import S3PrefixSensor  # noqa
 
 warnings.warn(
     "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.s3`.",
     DeprecationWarning,
-    stacklevel=2
+    stacklevel=2,
 )
